+++ src/Products/Sessions/__init__.py
@@ -10,19 +10,15 @@
 # FOR A PARTICULAR PURPOSE
 #
 ##############################################################################
+""" Session managemnt product initialization
 """
-Session initialization routines
-
-$Id$
-"""
-
-import ZODB # this is for testrunner to be happy
-import BrowserIdManager
-import SessionDataManager
-from BrowserIdManager import BrowserIdManagerErr
-from SessionDataManager import SessionDataManagerErr
+from Products.Sessions.interfaces import BrowserIdManagerErr    #BBB
+from Products.Sessions.interfaces import SessionDataManagerErr  #BBB
 
 def initialize(context):
+    import BrowserIdManager
+    import SessionDataManager
+
     context.registerClass(
         BrowserIdManager.BrowserIdManager,
         icon="www/idmgr.gif",
@@ -41,6 +37,7 @@ def initialize(context):
 
     context.registerHelp()
     context.registerHelpTitle("Zope Help")
+
     # do module security declarations so folks can use some of the
     # module-level stuff in PythonScripts
     #
@@ -51,8 +48,13 @@ def initialize(context):
     security = ModuleSecurityInfo('Products')
     security.declarePublic('Sessions')
     security.declarePublic('Transience')
+    security = ModuleSecurityInfo('Products.Sessions.interfaces')
+    security.declareObjectPublic()
+    security.setDefaultAccess('allow')
+    security = ModuleSecurityInfo('Products.Transience')
+    security.declarePublic('MaxTransientObjectsExceeded')
+
+    #BBB for names which should be imported from Products.Sessions.interfaces
     security = ModuleSecurityInfo('Products.Sessions')
     security.declarePublic('BrowserIdManagerErr')
     security.declarePublic('SessionDataManagerErr')
-    security = ModuleSecurityInfo('Products.Transience')
-    security.declarePublic('MaxTransientObjectsExceeded')